+++ personalityquiz.php
@@ -1,0 +1,41 @@
+<!DOCTYPE html>
+<html>
+<head>
+
+	<meta charset="UTF-8">
+	<meta name="description" content="Sweet Elite is a dating sims inspired by the popular Japanese Otome.Flirt with students and uncover Arlington Academy's dark secrets.Choose your own answers!">
+	
+	<title>Sweet Elite: Flirt and Uncover the Secrets of Arlington Academy!<</title>
+
+	<link href="css/se-stylesheet.css" rel="stylesheet" type="text/css">
+	<link href='https://fonts.googleapis.com/css?family=Roboto:400,100,100italic,300,300italic,400italic,500,500italic,700,700italic,900,900italic' rel='stylesheet' type='text/css'>
+	<link href="https://fonts.googleapis.com/css?family=Great+Vibes" rel="stylesheet" type='text/css'>
+
+
+	<link rel="icon" type="image/ico" href="images/favicon.png"/>
+	<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+
+
+</head>
+
+<body>
+
+	<div id="popupHandler"></div>
+
+	<?php include_once 'partials/headers.php' ?>
+
+	<div class = "main_content">
+
+		<div id = "quiz" class="card-firstdiv"></div>
+		<div id = "quizCard" class="card" style="background-color: white; border-radius: 10px;"></div>
+
+	</div>
+
+	<?php include_once 'partials/footers.php' ?>
+
+</body>
+
+<script src="js/personalityQuizQuestions.js"></script>
+<script src="js/personalityQuiz.js"></script>
+
+</html>